Implement /files endpoint
Note that we no longer convert the body to a String at any stage. The
Display implementation will no longer show the body if used to print.
Instead, we receive the body as an octet stream, store it, and send it
back in the same form. The reason for this, is to support sending different
binaries and file types without losing data due to conversion.